examples/citrineos/requirements.txt
requests>=2.31
pika>=1.3,<2.0
python-dotenv>=1.0
